+++ QuickFIXn/SessionFactory.cs
@@ -162,6 +162,10 @@ internal class SessionFactory
             session.RequiresOrigSendingTime = settings.GetBool(SessionSettings.RESETSEQUENCE_MESSAGE_REQUIRES_ORIGSENDINGTIME);
         if (settings.Has(SessionSettings.CME_ENHANCED_RESEND))
             session.CmeEnhancedResend = settings.GetBool(SessionSettings.CME_ENHANCED_RESEND);
+        if (settings.Has(SessionSettings.REDACT_FIELDS_IN_LOGS))
+            session.RedactFieldsInLogs = settings.GetIntArray(SessionSettings.REDACT_FIELDS_IN_LOGS);
+        if (settings.Has(SessionSettings.REDACTION_LOG_TEXT))
+            session.RedactionLogText = settings.GetString(SessionSettings.REDACTION_LOG_TEXT);
 
         return session;
     }