+++ QuickFIXn/SessionSettings.cs
@@ -73,6 +73,8 @@ public class SessionSettings
     public const string SOCKET_IGNORE_PROXY = "SocketIgnoreProxy";
     public const string ENCODING = "Encoding";
     public const string CME_ENHANCED_RESEND = "CmeEnhancedResend";
+    public const string REDACT_FIELDS_IN_LOGS = "RedactFieldsInLogs";
+    public const string REDACTION_LOG_TEXT = "RedactionLogText";
 
     public const string SSL_ENABLE = "SSLEnable";
     public const string SSL_SERVERNAME = "SSLServerName";