+++ UnitTests/SettingsDictionaryTests.cs
@@ -21,8 +21,11 @@ public class SettingsDictionaryTests
         d.SetString("STRINGKEY2", "stringvalue2");
         Assert.That(d.GetString("STRINGKEY1"), Is.EqualTo("STRINGVALUE1"));
         Assert.That(d.GetString("STRINGKEY2"), Is.EqualTo("stringvalue2"));
-        Assert.That(d.GetString("STRINGKEY2", true), Is.EqualTo("STRINGVALUE2"));
         Assert.Throws<ConfigError>(delegate { d.GetString("STRINGKEY3"); });
+
+        #pragma warning disable CS0618
+        Assert.That(d.GetString("STRINGKEY2", true), Is.EqualTo("STRINGVALUE2"));
+        #pragma warning restore CS0618
     }
 
     [Test]
@@ -187,5 +190,39 @@ public class SettingsDictionaryTests
         Assert.That(dupe.Count, Is.EqualTo(2));
         Assert.That(dupe.GetString("uNo"), Is.EqualTo("One"));
         Assert.That(dupe.GetString("DOs"), Is.EqualTo("2"));
+    }
+
+    [Test]
+    public void TestGetIntArray() {
+        SettingsDictionary settings = new("test");
+
+        // empty string becomes empty array
+        settings.SetString("myintarray", "");
+        Assert.That(settings.GetIntArray("myintarray"), Is.Empty);
+
+        int[] expected = [1, 33, 999];
+
+        // the intended input format
+        settings.SetString("myintarray", "1,33,999");
+        Assert.That(settings.GetIntArray("myintarray"), Is.EqualTo(expected));
+
+        // trailing & repeated commas ignored
+        settings.SetString("myintarray", "1,33,,999,");
+        Assert.That(settings.GetIntArray("myintarray"), Is.EqualTo(expected));
+
+        // spaces ignored
+        settings.SetString("myintarray", "1, 33, , 999, ");
+        Assert.That(settings.GetIntArray("myintarray"), Is.EqualTo(expected));
+
+        // sorted, dupes consolidated
+        settings.SetString("myintarray", "33, 1, , 999, 33, 1");
+        Assert.That(settings.GetIntArray("myintarray"), Is.EqualTo(expected));
+
+        // key not found
+        Assert.Throws<ConfigError>(delegate { settings.GetIntArray("nope"); });
+
+        // invalid values
+        settings.SetString("myintarray", "1, 33, , 999, fourteen, ");
+        Assert.Throws<ConfigError>(delegate { settings.GetIntArray("myintarray"); });
     }
 }