+++ RELEASE_NOTES.md
@@ -35,6 +35,7 @@ What's New
 * #961 - new GenerateKeys app to create Example-app SSL certs (dckorben/gbirchmeier)
 * #1001 - AcceptanceTests bug: double.Parse with InvariantCulture (gbirchmeier)
 * #562 - deprecate Message.IsHeaderField without transport DD param (gbirchmeier)
+* #949 - new settings RedactFieldsInLogs & RedactionLogText (gbirchmeier)
 
 
 ### v1.14.0